User getById only form service, not repo + small fixes

+++ DeedCurrencyPay/Services/IUserService.cs
@@ -4,6 +4,6 @@ namespace DeedCurrencyPay.API.Services
 {
     public interface IUserService
     {
-        User GetById(int id);
+        User GetById(long id);
     }
 }
